pyproject.toml
[tool.poetry]
name = "chatbot"
version = "0.1.0"
description = ""
authors = ["Yannick Schwarz"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
llama-index = "*"
python-dotenv = ">=1.1.1,<2.0.0"
uvicorn = ">=0.35.0,<0.36.0"
fastapi = ">=0.115.14,<0.116.0"
azure-identity = ">=1.23.0,<2.0.0"
azure-keyvault-secrets = ">=4.10.0,<5.0.0"
azure-storage-blob = ">=12.25.1,<13.0.0"
openai = ">=1.93.0,<2.0.0"
qdrant-client = ">=1.9.0,<2.0.0"
llama-index-vector-stores-qdrant = "*"
langchain = ">=0.3.26,<0.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
